requirements.txt
certifi==2017.7.27.1
Django==1.11.6
django-filter==1.1.0
djangorestframework==3.7.1
Markdown==2.6.9
pytz==2017.2
